Support string method calls
Technically, to be 100% compatible, this should be calling through
a metatable assigned to all strings which has __index pointing to
the string global table. This intermediate approach works for all
sane use cases without requiring so much work. (It is not possible
to *set* a metatable on a string in PUC-Lua, but it is at least
possible to call `getmetatable` on a string and then mess with it.)